backend/requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2024.7.4
click==8.1.7
cloudinary==1.40.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
gunicorn==23.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.2
whitenoise==6.7.0
